# Bounds checks elimination