@klauspost 2024-06-17
If you can get `m.R.Next(...)` and `m.R.Peek(...)` to inline the "happy path" that would probably give a similar speedup.

Tried it briefly, but even...

```Go
func (r *Reader) Next(n int) ([]byte, error) {
	// Have happy path be short and possible to inline.
	if len(r.data)-r.n >= n && r.state == nil {
		return r.data[r.n:], nil
	}
	return r.next(n)
}

// next is a helper function for Next to be called when the buffer does not contain n entries.
func (r *Reader) next(n int) ([]byte, error) {
...
```
...is too costly.

Also `m.R.Peek(1)` is so common that could have a specialized call `func (r *Reader) PeekByte() (byte, error)` - but again I couldn't get it to inline with some basic attempts. 

It was slightly faster by having the "happy path" first, but nothing major. Maybe like 5%.