@mhr3 2024-06-17
> Awesome find!
> 
> ```
> func getMint64(b []byte) int64 {
> 	return int64(binary.BigEndian.Uint64(b[1:9]))
> }
> ```
> 
> ... is a shorter version with only one check (spits out same code on amd64) - which is also inlined.
> 
> No need to change

I did try this, but from my benchmarking this was teeny bit slower (on arm64), did not check the generated assembly though.